xds: Add v3 implementation
